+++ frontend/src/App.js
@@ -3,9 +3,9 @@ import "./App.css";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Home from "./pages/Home/Home.js";
 import Navbar from "./components/Navbar/Navbar";
-import Iceland from "./pages/Iceland/Iceland";
-import Worldwide from "./pages/Worldwide/Worldwide";
 import About from "./pages/About/About";
+import FindBestThisWeek from "./pages/FindBestThisWeek/FindBestThisWeek";
+import FindBestToday from "./pages/FindBestToday/FindBestToday";
 
 const App = () => {
   return (
@@ -14,8 +14,12 @@ const App = () => {
         <Navbar />
         <Switch>
           <Route exact path="/" render={() => <Home />} />
-          <Route exact path="/iceland" render={() => <Iceland />} />
-          <Route exact path="/worldwide" render={() => <Worldwide />} />
+          <Route exact path="/bestToday" render={() => <FindBestToday />} />
+          <Route
+            exact
+            path="/bestThisWeek"
+            render={() => <FindBestThisWeek />}
+          />
           <Route exact path="/about" render={() => <About />} />
         </Switch>
       </Router>
