+++ frontend/src/pages/FindBestToday/FindBestToday.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const FindBestToday = () => {
+  return <>Hvar er besta veðrið í dag?</>;
+};
+
+export default FindBestToday;
