--- frontend/src/pages/Iceland/Iceland.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Iceland = () => {
-  return <p>Ski resorts in Iceland</p>;
-};
-
-export default Iceland;
