--- frontend/src/pages/Worldwide/Worldwide.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Worldwide = () => {
-  return <p>Worldwide ski resorts</p>;
-};
-
-export default Worldwide;
